% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.n2.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Function to draw a heatmap based on prepared "display list" (dl)}
\usage{
draw_heatmap(dl, set.oma = TRUE, ...)
}
\arguments{
\item{dl}{(\code{list}). The "display list": A list with all parameters needed for plotting the heatmap.}

\item{set.oma}{(\code{logical}). Should the outer margins be set to optimized values? Defaults to \code{TRUE}.}

\item{...}{(\code{list}). Additional parameters passed to \code{\link[base]{plot}}.}
}
\value{
Invisibly returns a list with graphical parameters. But mainly the function is called for its side-effect
    of generating a plot.
}
\description{
Normally, \code{draw_heatmap} is called from within \code{\link[heatmapFlex]{heatmap.n2}} (if argument
    \code{plot} is \code{TRUE} which is the default) but if that is not desired, e.g., for zooming, it can be called
    separately.
}
\seealso{
\code{\link[graphics]{par}}
}
