% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.n2.R
\docType{package}
\name{heatmapS}
\alias{heatmapS}
\title{Tools to generate zoomable and splittable heatmaps}
\description{
The package has a number of tools supporting more flexible heatmaps. The graphics is grid-like using
 the old graphics system. The main function is \code{heatmap.n2}, which is a wrapper around the various functions
 constructing individual parts of the heatmap, like sidebars, picket plots, legends etc. The function supports zooming
 and splitting, i.e., having (unlimited) small heatmaps underneath each other in one plot deriving from the same data set,
 e.g., clustered and ordered by a supervised clustering method.
}
\details{
\tabular{ll}{
Package: \tab heatmapS\cr
Type: \tab Package\cr
Initial version: \tab 0.1-0\cr
Created: \tab 2021-08-11\cr
License: \tab File LICENSE\cr
LazyLoad: \tab yes\cr
}
}
\author{
Vidal Fey <vidal.fey@gmail.com>, Henri Sara <henri.sara@gmail.com>
Maintainer: Vidal Fey <vidal.fey@gmail.com>
}
\keyword{package}
