% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.n2.R
\name{zoom_heatmap}
\alias{zoom_heatmap}
\title{Function to zoom into an already drawn heatmap}
\usage{
zoom_heatmap(dl, ...)
}
\arguments{
\item{dl}{(\code{list}). The "display list": A list with all parameters needed for plotting the heatmap.}

\item{...}{(\code{list}). Additional parameters passed to \code{\link[heatmapFlex]{draw_heatmap}}
(and eventually to \code{\link[base]{plot}}).}
}
\value{
Invisibly returns a list with graphical parameters but is mainly called for its side-effect of drawing a zoomed plot.
}
\description{
After the heatmap was drawn in a screen device such as X11, windows and quartz it can be zoomed
    by using the mouse.
}
\seealso{
\code{\link[graphics]{locator}}
}
